add 2 statistics

+++ CurrencyConverterStatisticsService/src/main/java/ru/microservices/currency/converters/currencyconverterstatisticsservice/api/ConversionResponse.java
@@ -3,7 +3,7 @@ package ru.microservices.currency.converters.currencyconverterstatisticsservice.
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -12,5 +12,5 @@ public class ConversionResponse {
     private final String from;
     private final String to;
     private final long count;
-    private final BigInteger amount;
+    private final BigDecimal amount;
 }
